requirements.txt
streamlit==1.54.0
crewai==1.9.3
crewai-tools==1.9.3
python-pptx==1.0.2
PyPDF2==3.0.1
python-dotenv
